Full filter and sorting functionality
All requirements for filtering and sorting have been implemented
For the sorting, I chose to add an option to sort by ID

Modifying JSON data, to have better example data.

At the moment, all the code is in the main.js file. I will try to clean up in one of the next commits.